[dtraj] making full dtraj api available
