+++ .gitignore
@@ -1,3 +1,4 @@
+# compilation / build files
 *.orig
 *.bak
 build*
@@ -15,17 +16,24 @@ __pycache__
 /pylint/pylint_*.txt
 /dist
 .externalToolBuilders
+
 # ignore untitled ipython notebooks
 Untitled*.ipynb
+
 # ignore c-files generated by cython
 /*/estimation/dense/mle_trev_given_pi.c
 /*/estimation/sparse/mle_trev.c
 /*/estimation/sparse/mle_trev_given_pi.c
 /*/estimation/dense/sampler_rev.c
 /*/estimation/dense/sampler_revpi.c
+
 # generated from MANIFEST.in
 MANIFEST
 doc/source/api/generated/
 .ipynb_checkpoints
 .project
 .pydevproject
+
+# project files
+.idea
+
