+++ msmtools/dtraj/__init__.py
@@ -46,6 +46,27 @@ Discrete trajectory io
    save_discrete_trajectory -  write microstate trajectory to binary file
    save_dtraj
 
+Simple statistics
+=================
+
+.. autosummary::
+   :toctree: generated/
+
+   count_states
+   visited_set
+   number_of_states
+   index_states
+
+Sampling trajectory indexes
+===========================
+
+.. autosummary::
+   :toctree: generated/
+
+   sample_indexes_by_distribution
+   sample_indexes_by_state
+   sample_indexes_by_sequence
+
 """
 from __future__ import absolute_import
 from .api import *
